+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.6.2] - 2025-07-10
 
+### Added
+
+- Exception: when SourceType is 'AzureDatabricksDeltaLakeSource' and SinkType is 'DelimitedTextSink', staging via 'SinkStagingSettings' must be enabled
+
 ### Changed
 
 - allowed SourceType 'AzureDatabricksDeltaLakeSink' to 'AzureDatabricksDeltaLakeSource' in 'New-CopyActivity'